utils/synbot.py
from dotenv import load_dotenv
import cohere
import os

load_dotenv()
api_key = os.getenv('COHERE_API_KEY')
co = cohere.Client(api_key)

def get_budget_insights(user_query, transactions_text):
    prompt = f"""User query: {user_query}\nTransactions list: {transactions_text}\n
    You are SynBot, a financial AI assistant developed by Maria Asif for the Syntego Finance Tracker. 
    Your job is **ONLY** to assist users with their **financial queries** including budgeting, expense tracking, and savings suggestions.
    If the user asks unrelated questions (e.g., about weather, jokes, or anything personal), reply with: 
    "I can only assist with finacial-related questions. Please ask me something about your finace."
    If user asks about making changes in expenses or income to delete or add, simply respond:
    "I can assist you with analyzing or summarizing your finances, but you’ll need to use the app interface to make changes."
    If the user asks about  **yourself**, simply respond:
    "I am SynBot, a finacial asssitant built by Maria Asif to help with budegting and expense management."""

    response = co.chat(
    model='command-r-plus',
    message=prompt,
    temperature=0.7,
    max_tokens=200
)


    # return the response from Cohere API
    return response.text
