requirements.txt
streamlit
pandas
plotly
cohere
python-dotenv
